+++ src/main/java/com/kuit/healthmate/domain/user/User.java
@@ -1,4 +1,4 @@
-package com.kuit.healthmate.domain;
+package com.kuit.healthmate.domain.user;
 
 
 import com.kuit.healthmate.domain.supplement.Supplement;
